Simplify Decimal constructor
As fromNumber, fromString and fromDecimal are guaranteed to set sign,
mag and layer, the initial values of sign, mag and layer (NaNs) are
effectively unused.

As TypeScript compiles down default values to "immediately set default
values at the start of the constructor", this commit replaces the
default values of NaNs to 0s, which allows us to remove the
`value === undefined` branch of the constructor.